% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_summary.R
\name{po_counts}
\alias{po_counts}
\title{pocount is available from e.g. 'apt install translate-toolkit'}
\usage{
po_counts(files)
}
\description{
pocount is available from e.g. 'apt install translate-toolkit'
}
